[v3-0-test] Move example_dags in standard provider to example_dags in sources (#51260)
Folow up after #51192.

Actually choosing the same approach as for system tests was
pretty wrong and even before attempting to have final solution
for example_dags/system tests we should move the standard examples
out of "tests.system" as rightfully pointed out by Jed in
the #51192.

The snippets of code to run the tests as pytests are not neeeded
and they are actually even harmful as they cause import errors
when no tests_common is around
(cherry picked from commit cc29e027ce777452c214187e675eb247a6e231ad)
