+++ airflow-core/src/airflow/example_dags/standard
@@ -1,1 +1,1 @@
-../../../../providers/standard/tests/system/standard/
+../../../../providers/standard/src/airflow/providers/standard/example_dags
