+++ devel-common/src/docs/provider_conf.py
@@ -61,7 +61,6 @@ from docs.utils.conf_constants import (
     get_html_theme_options,
     get_intersphinx_mapping,
     get_rst_epilogue,
-    get_rst_filepath_from_path,
 )
 from sphinx_exts.provider_yaml_utils import load_package_data
 
@@ -161,10 +160,10 @@ exclude_patterns = [
     *[f"_api/tests/system/{subpackage}/index.rst" for subpackage in empty_subpackages],
 ]
 
-exclude_patterns.extend(
-    get_rst_filepath_from_path(f, AIRFLOW_REPO_ROOT_PATH)
-    for f in BASE_PROVIDER_SRC_PATH.rglob("example_dags")
-)
+# exclude_patterns.extend(
+#     get_rst_filepath_from_path(f, AIRFLOW_REPO_ROOT_PATH)
+#     for f in BASE_PROVIDER_SRC_PATH.rglob("example_dags")
+# )
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["templates"]
@@ -296,7 +295,6 @@ autoapi_ignore.extend(
     (
         "*/airflow/__init__.py",
         "*/airflow/providers/__init__.py",
-        "*/example_dags/*",
         "*/providers/__init__.py",
         "*/conf/*",
     )
