+++ docs/spelling_wordlist.txt
@@ -1311,6 +1311,7 @@ pluggable
 pluggy
 plyvel
 png
+PoC
 PodManager
 podName
 PodSpec
