+++ providers/standard/docs/sensors/bash.rst
@@ -25,7 +25,7 @@ BashSensor
 Use the :class:`~airflow.providers.standard.sensors.bash.BashSensor` to use arbitrary command for sensing. The command
 should return 0 when it succeeds, any other value otherwise.
 
-.. exampleinclude:: /../tests/system/standard/example_sensors.py
+.. exampleinclude:: /../src/airflow/providers/standard/example_dags/example_sensors.py
     :language: python
     :dedent: 4
     :start-after: [START example_bash_sensors]
