+++ providers/standard/docs/sensors/python.rst
@@ -34,7 +34,7 @@ value to be True.
     .. tab-item:: @task.sensor
         :sync: taskflow
 
-        .. exampleinclude:: /../tests/system/standard/example_sensor_decorator.py
+        .. exampleinclude:: /../src/airflow/providers/standard/example_dags/example_sensor_decorator.py
             :language: python
             :dedent: 4
             :start-after: [START wait_function]
@@ -43,7 +43,7 @@ value to be True.
     .. tab-item:: PythonSensor
         :sync: operator
 
-        .. exampleinclude:: /../tests/system/standard/example_sensors.py
+        .. exampleinclude:: /../src/airflow/providers/standard/example_dags/example_sensors.py
             :language: python
             :dedent: 4
             :start-after: [START example_python_sensors]
