+++ providers/standard/src/airflow/providers/standard/example_dags/example_bash_decorator.py
@@ -112,9 +112,3 @@ def example_bash_decorator():
 
 
 example_bash_decorator()
-
-
-from tests_common.test_utils.system_tests import get_test_run  # noqa: E402
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(dag)
