+++ providers/standard/src/airflow/providers/standard/example_dags/example_branch_operator_decorator.py
@@ -140,9 +140,3 @@ with DAG(
 
         # Label is optional here, but it can help identify more complex branches
         random_choice_venv >> Label(option) >> t >> join_venv
-
-
-from tests_common.test_utils.system_tests import get_test_run  # noqa: E402
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(dag)
