+++ providers/standard/src/airflow/providers/standard/example_dags/example_external_task_marker_dag.py
@@ -96,9 +96,3 @@ with DAG(
 
     child_task3 = EmptyOperator(task_id="child_task3")
     child_task1 >> child_task2 >> child_task3
-
-
-from tests_common.test_utils.system_tests import get_test_run  # noqa: E402
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(parent_dag)
