+++ providers/standard/src/airflow/providers/standard/example_dags/example_latest_only.py
@@ -1,3 +1,4 @@
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,27 +15,26 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""Example of the LatestOnlyOperator"""
+
 from __future__ import annotations
 
-from datetime import datetime
+import datetime
 
-from airflow import DAG
-from airflow.providers.standard.operators.bash import BashOperator
+from airflow.providers.standard.operators.empty import EmptyOperator
+from airflow.providers.standard.operators.latest_only import LatestOnlyOperator
+from airflow.sdk import DAG
 
 with DAG(
-    dag_id="child_dag",
-    start_date=datetime(2022, 1, 1),
-    schedule="@once",
+    dag_id="latest_only",
+    schedule=datetime.timedelta(hours=4),
+    start_date=datetime.datetime(2021, 1, 1),
     catchup=False,
-    tags=["example", "async", "core"],
+    tags=["example2", "example3"],
 ) as dag:
-    BashOperator(
-        task_id="child_task",
-        bash_command="echo 1; sleep 1; echo 2; sleep 2; echo 3; sleep 3",
-    )
+    # [START howto_operator_latest_only]
+    latest_only = LatestOnlyOperator(task_id="latest_only")
+    # [END howto_operator_latest_only]
+    task1 = EmptyOperator(task_id="task1")
 
-
-from tests_common.test_utils.system_tests import get_test_run
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(dag)
+    latest_only >> task1
