+++ providers/standard/src/airflow/providers/standard/example_dags/example_python_decorator.py
@@ -130,9 +130,3 @@ def example_python_decorator():
 
 
 example_dag = example_python_decorator()
-
-
-from tests_common.test_utils.system_tests import get_test_run  # noqa: E402
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(example_dag)
