+++ providers/standard/src/airflow/providers/standard/example_dags/example_python_operator.py
@@ -145,9 +145,3 @@ with DAG(
     # [END howto_operator_external_python]
 
     run_this >> external_python_task >> virtualenv_task
-
-
-from tests_common.test_utils.system_tests import get_test_run  # noqa: E402
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(dag)
