+++ providers/standard/src/airflow/providers/standard/example_dags/example_sensor_decorator.py
@@ -64,9 +64,3 @@ tutorial_etl_dag = example_sensor_decorator()
 # [END dag_invocation]
 
 # [END tutorial]
-
-
-from tests_common.test_utils.system_tests import get_test_run  # noqa: E402
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(dag)
