+++ providers/standard/src/airflow/providers/standard/example_dags/example_sensors.py
@@ -130,9 +130,3 @@ with DAG(
     t8 >> tx
     [t9, t10] >> tx
     t11 >> tx
-
-
-from tests_common.test_utils.system_tests import get_test_run  # noqa: E402
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(dag)
