+++ providers/standard/src/airflow/providers/standard/example_dags/example_trigger_controller_dag.py
@@ -44,9 +44,3 @@ with DAG(
     )
 
     # [END howto_operator_trigger_dagrun]
-
-
-from tests_common.test_utils.system_tests import get_test_run
-
-# Needed to run the example DAG with pytest (see: tests/system/README.md#run_via_pytest)
-test_run = get_test_run(dag)
